Document comparision module added

+++ src/document_compare/data_ingestion.py
@@ -1,6 +1,6 @@
 import sys
 import os
-import fitz
+import fitz #type: ignore
 from pathlib import Path
 from logger.custom_logger import CustomLogger
 from exception.custom_exception import DocumentPortalException
@@ -10,9 +10,9 @@ class DocumentIngestion:
     
     """
 
-    def __init__(self, base_dir: str):
+    def __init__(self, base_dir: str = "data\\document_compare"):
         self.log = CustomLogger().get_logger(__name__)
-        self.base_dir = base_dir
+        self.base_dir = Path(base_dir)
         self.base_dir.mkdir(parents=True, exist_ok=True)
 
     def delete_existing_file(self):
@@ -20,7 +20,7 @@ class DocumentIngestion:
         deletes the existing file in the base directory.
         """
         try:
-            is self.base_dir.exists() and self.base_dir.is_dir():
+            if self.base_dir.exists() and self.base_dir.is_dir():
                 for file in self.base_dir.iterdir():
                     if file.is_file():
                         file.unlink()
@@ -30,7 +30,7 @@ class DocumentIngestion:
             self.log.error(f"Error occurred while deleting the file: {e}")
             raise DocumentPortalException("Failed to delete the file", sys)
 
-    def save_uploaded_file(self, reference_file, actual_file):
+    def save_uploaded_files(self, reference_file, actual_file):
         """
         saves the uploaded file to the base directory.
         """
@@ -40,10 +40,10 @@ class DocumentIngestion:
             # Save the new file logic here
             reference_path = self.base_dir / reference_file.name
             actual_path = self.base_dir / actual_file.name
-
-            if not in reference_file.name.endswith('.pdf') or actual_file.name.endswith('.pdf'):
+            
+            if not (reference_file.name.endswith('.pdf') and actual_file.name.endswith('.pdf')):
                 raise ValueError("Only PDF files are supported.")
-
+            
             with open(reference_path, 'wb') as f:
                 f.write(reference_file.getbuffer())
 
@@ -53,10 +53,12 @@ class DocumentIngestion:
             self.log.info(f"Files saved successfully", reference_file=str(reference_path), actual_file=str(actual_path))
 
             return reference_path, actual_path
+        
         except Exception as e:
             self.log.error(f"Error occurred while saving the file: {e}")
             raise DocumentPortalException("Failed to save the file", sys)
         
+
     def read_pdf(self, pdf_path: Path) -> str:
         """
         Gets the the PDF file and extract text from it.
@@ -72,9 +74,33 @@ class DocumentIngestion:
                     if text.strip():
                         all_text.append(f"\n --- Page{page_num + 1} --- \n")
 
-                log.info(f"Extracted text successfully", file=str(pdf_path), pages=len(all_text))
+                self.log.info(f"Extracted text successfully", file=str(pdf_path), pages=len(all_text))
                 return "\n".join(all_text)
 
         except Exception as e:
             self.log.error(f"Error occurred while reading PDF file: {e}")
             raise DocumentPortalException("Failed to read the pdf", sys) 
+        
+    def combine_documents(self) -> str:
+        """
+        Combines the text from the two PDF files.
+        
+        """
+        try:
+            content_dict = {}
+            doc_parts = []
+
+            for filename in sorted(self.base_dir.iterdir()):
+                if filename.is_file() and filename.suffix == '.pdf':
+                    content_dict[filename.name] = self.read_pdf(filename)
+            
+            for filename, content in content_dict.items():
+                doc_parts.append(f"Document: {filename}\n{content}")
+
+            combined_text = f"\n\n".join(doc_parts)
+            self.log.info("Documents combined successfully", count = len(doc_parts))
+            return combined_text
+        
+        except Exception as e:
+            self.log.error(f"Error occurred while combining documents: {e}")
+            raise DocumentPortalException("Failed to combine the documents", sys)
